+++ src/libraries/Common/src/Extensions/ParameterDefaultValue/ParameterDefaultValue.netcoreapp.cs
@@ -10,48 +10,16 @@ using System.Runtime.CompilerServices;
 
 namespace Microsoft.Extensions.Internal
 {
-    internal static class ParameterDefaultValue
+    internal static partial class ParameterDefaultValue
     {
-        public static bool TryGetDefaultValue(ParameterInfo parameter, out object? defaultValue)
+        public static bool CheckHasDefaultValue(ParameterInfo parameter)
         {
-            bool tryToGetDefaultValue = true;
-            defaultValue = null;
-            bool hasDefaultValue = parameter.HasDefaultValue;
-
-            if (hasDefaultValue)
-            {
-                if (tryToGetDefaultValue)
-                {
-                    defaultValue = parameter.DefaultValue;
-                }
-
-                bool isNullableParameterType = parameter.ParameterType.IsGenericType &&
-                    parameter.ParameterType.GetGenericTypeDefinition() == typeof(Nullable<>);
-
-                // Workaround for https://github.com/dotnet/runtime/issues/18599
-                if (defaultValue == null && parameter.ParameterType.IsValueType
-                    && !isNullableParameterType) // Nullable types should be left null
-                {
-                    defaultValue = CreateValueType(parameter.ParameterType);
-                }
-
-                [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2067:UnrecognizedReflectionPattern",
-                    Justification = "CreateValueType is only called on a ValueType. You can always create an instance of a ValueType.")]
-                static object? CreateValueType(Type t) =>
-                    RuntimeHelpers.GetUninitializedObject(t);
-
-                // Handle nullable enums
-                if (defaultValue != null && isNullableParameterType)
-                {
-                    Type? underlyingType = Nullable.GetUnderlyingType(parameter.ParameterType);
-                    if (underlyingType != null && underlyingType.IsEnum)
-                    {
-                        defaultValue = Enum.ToObject(underlyingType, defaultValue);
-                    }
-                }
-            }
-
-            return hasDefaultValue;
+            return parameter.HasDefaultValue;
         }
+
+        [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2067:UnrecognizedReflectionPattern",
+            Justification = "CreateValueType is only called on a ValueType. You can always create an instance of a ValueType.")]
+        internal static object? CreateValueType(Type t) =>
+            RuntimeHelpers.GetUninitializedObject(t);
     }
 }
